Introduced Auth::writeSession
Introduced the new method Auth::writeSession. It does the same Auth::login
did before, but as the method is needed in update, the name does not make
sense. Also, Auth::login will also identify the user based on the param it
receives, so it's better to separate concerns. Auth::update now receives
the user object instead of the data, so the developer is responsible for
providing the right info (and we use one less db query)